requirements.txt
web3>=6.0.0
python-dotenv==1.0.0
eth-abi>=4.0.0
eth-account>=0.8.0
eth-typing>=3.0.0
eth-utils>=2.1.0
requests>=2.28.0
python-dateutil==2.8.2 